+++ CHANGELOG.md
@@ -7,6 +7,7 @@ As of 12/13/2022, this repo has been renamed from "nftc-open-contracts" to "nftc
 ## Version -- 1.5.next [Not published]
 - Update license for 2023
 - update to typescript 5.2.2
+- cleanups to nftspecchecker test
 
 
 ## Version -- 1.5.6